requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.5
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
ciscoconfparse==1.4.11
colorama==0.4.3
cryptography==2.4.2
dnspython==1.16.0
future==0.17.1
idna==2.8
Jinja2==2.10.3
junos-eznc==2.2.0
lxml==4.2.5
MarkupSafe==1.1.0
napalm==2.3.3
ncclient==0.6.3
netaddr==0.7.19
netmiko==2.3.0
paramiko==2.4.2
passlib==1.7.2
pkg-resources==0.0.0
pyasn1==0.4.4
pycparser==2.19
pyeapi==0.8.2
pyIOSXR==0.53
PyNaCl==1.3.0
pynetbox==4.2.2
pynxos==0.0.3
pyserial==3.4
PyYAML==3.13
requests==2.21.0
scp==0.13.0
selectors2==2.0.1
six==1.12.0
textfsm==0.4.1
urllib3==1.24.2
